fixed warnings in msvc removed authorship for code that wasn't written by me

+++ code/trunk/src/orxonox/worldentities/ControllableEntity.h
@@ -165,7 +165,7 @@ namespace orxonox
 
 	    inline void setTeam(int team)
 	        { this->team_ = team; }
-	    inline float getTeam() const
+	    inline int getTeam() const
 	        { return this->team_; }
 
         protected:
